+++ src/maps/cc-map.js
@@ -6,7 +6,6 @@ import { WORLD_GEOJSON } from '../assets/world-110m.geo.js';
 import { dispatchCustomEvent } from '../lib/events.js';
 import { i18n } from '../lib/i18n.js';
 import * as leaflet from '../lib/leaflet-esm.js';
-import { HeatLayer } from '../lib/leaflet-heat.js';
 import { withResizeObserver } from '../mixins/with-resize-observer.js';
 import { leafletStyles } from '../styles/leaflet.js';
 
@@ -131,7 +130,7 @@ export class CcMap extends withResizeObserver(LitElement) {
 
     this._heatLayer
       .clearLayers()
-      .addLayer(new HeatLayer(heatPoints, heatOptions));
+      .addLayer(new leaflet.HeatLayer(heatPoints, heatOptions));
   }
 
   _updatePoints (newPoints) {